fix - aws/rds change port type

+++ aws/rds/main.go
@@ -17,7 +17,7 @@ type Config struct {
 	Status     string   `json:"status"`     // New instance status.
 	Endpoint   string   `json:"endpoint"`   // New instance endpoint.
 	Protection bool     `json:"protection"` // New instance is set to deletion protection.
-	Port       int64    `json:"port"`       // New instance port of endpoint.
+	Port       uint16   `json:"port"`       // New instance port of endpoint.
 	Zone       string   `json:"zone"`       // New instance Availability Zone.
 }
 
@@ -82,7 +82,7 @@ func (c *Config) Describe() {
 
 	if result.DBInstances[0].Endpoint != nil {
 		c.Endpoint = *result.DBInstances[0].Endpoint.Address
-		c.Port = *result.DBInstances[0].Endpoint.Port
+		c.Port = uint16(*result.DBInstances[0].Endpoint.Port)
 	}
 }
